# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fortress Platform API
  version: 1.0.0
  description: Public API for trusted, versioned Islamic knowledge data.
  x-platform-version: 0.16.0
servers:
  - url: https://api.fortressofmuslim.org/v1
    description: Production
  - url: https://api-test.fortressofmuslim.org/v1
    description: Test
paths:
...
  /duas:
    get:
      operationId: listDuas
      summary: List dua summaries
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        '200':
          description: Paginated dua summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuaListResponse'
        '400':
          description: Invalid pagination parameters
  /duas/search:
    get:
      operationId: searchDuas
      summary: Search dua titles and all text segments
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 200
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Paginated matching dua summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuaListResponse'
        '400':
          description: Invalid query or pagination parameters
  /duas/random:
    get:
      operationId: getRandomDua
      summary: Retrieve one random complete dua
      responses:
        '200':
          description: Complete random dua
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuaResponse'
        '404':
          description: No published dua is available
  /duas/{id}:
    get:
      operationId: getDua
      summary: Retrieve a complete dua
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: dua.hisn.001
      responses:
        '200':
          description: Complete dua record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuaResponse'
        '404':
          description: Dua not found
  /duas/{id}/evidence:
    get:
      operationId: getDuaEvidence
      summary: Retrieve source provenance and editorial history for a dua
      description: Returns the evidence currently attached to the record. Empty or pending evidence is represented honestly and never inferred.
      parameters:
        - $ref: '#/components/parameters/DuaId'
      responses:
        '200':
          description: Traceable source, taxonomy, verification, and correction metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuaEvidenceResponse'
        '404':
          description: Dua not found
  /duas/{id}/parts:
    get:
      operationId: listDuaParts
      summary: Retrieve every ordered part of a dua
      parameters:
        - $ref: '#/components/parameters/DuaId'
      responses:
        '200':
          description: Ordered swipe-sized dua parts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DuaPart'
        '404':
          description: Dua not found
  /duas/{id}/parts/{position}:
    get:
      operationId: getDuaPart
      summary: Retrieve one ordered part of a dua
      parameters:
        - $ref: '#/components/parameters/DuaId'
        - in: path
          name: position
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: One swipe-sized dua part
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/DuaPart'
        '400':
          description: Invalid part position
        '404':
          description: Dua or part not found
  /collections:
    get:
      operationId: listCollections
      summary: List published dua and Hadith collections
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [dua, hadith]
      responses:
        '200':
          description: Collection summaries with record and hierarchy counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CollectionSummary'
  /hadith:
    get:
      operationId: listHadith
      summary: List published Hadith summaries
      parameters:
        - $ref: '#/components/parameters/Collection'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Paginated Hadith summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HadithListResponse'
  /hadith/search:
    get:
      operationId: searchHadith
      summary: Full-text search published Hadith
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 200
        - $ref: '#/components/parameters/Collection'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Ranked Hadith summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HadithListResponse'
  /hadith/{id}:
    get:
      operationId: getHadith
      summary: Retrieve one complete Hadith
      parameters:
        - in: path
          name: id
          required: true
          description: Fortress canonical record ID
          schema:
            type: string
          example: hadith.bukhari.1
      responses:
        '200':
          description: Complete Hadith, hierarchy, grading, text, and references
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HadithResponse'
        '404':
          description: Hadith not found
  /hadith/resolve:
    get:
      operationId: resolveCanonicalHadithPath
      summary: Resolve a sequential Fortress Hadith path
      description: Resolves the published canonical record represented by a Fortress URL such as /bukhari/book1/1.
      parameters:
        - in: query
          name: collection
          required: true
          schema: { type: string }
          example: bukhari
        - in: query
          name: book
          required: true
          schema: { type: string }
          example: '1'
        - in: query
          name: number
          required: true
          schema: { type: string }
          example: '1'
      responses:
        '200':
          description: Published canonical Hadith
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HadithResponse'
        '404':
          description: Published canonical record not found
  /ask:
    post:
      operationId: askFortress
      tags: [AI]
      summary: Ask a source-grounded question
      description: Retrieves semantically relevant published records and produces a concise answer with numbered source citations. This endpoint does not replace scholarly guidance.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskRequest'
      responses:
        '200':
          description: Grounded answer and traceable sources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskResponse'
        '400':
          description: Question is missing or invalid
        '429':
          description: Daily assistant allowance reached
  /queries/{id}:
    get:
      operationId: executeNamedQuery
      summary: Execute an owner-scoped named query
      description: Runs a saved, allowlisted read-only operation. Raw SQL is never accepted.
      security:
        - fortressApiKey: []
        - fortressOAuth:
            - content:read
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Named query result
        '404':
          description: Query not found or not owned by the credential principal
components:
...
  securitySchemes:
    fortressApiKey:
      type: apiKey
      in: header
      name: X-Fortress-API-Key
      description: Developer API key created in the Fortress Developer Portal.
    fortressOAuth:
      type: oauth2
      description: OAuth 2.1 access token issued by Fortress Platform.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.fortressofmuslim.org/api/auth/oauth2/authorize
          tokenUrl: https://auth.fortressofmuslim.org/api/auth/oauth2/token
          scopes:
            content:read: Read published dua records.
            content:search: Search published dua content.
            dataset:read: Read dataset provenance and publication metadata.
        clientCredentials:
          tokenUrl: https://auth.fortressofmuslim.org/api/auth/oauth2/token
          scopes:
            content:read: Read published dua records.
            content:search: Search published dua content.
            dataset:read: Read dataset provenance and publication metadata.
  parameters:
    DuaId:
      in: path
      name: id
      required: true
      description: Canonical or legacy dua ID
      schema:
        type: string
      example: dua.hisn.001
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      in: query
      name: cursor
      schema:
        type: string
    Collection:
      in: query
      name: collection
      description: Collection slug such as bukhari, muslim, or tirmidhi
      schema:
        type: string
  schemas:
...
    DuaPart:
      type: object
      required: [duaId, position, segmentCount, segments]
      properties:
        duaId:
          type: string
        position:
          type: integer
          minimum: 1
        segmentCount:
          type: integer
          minimum: 0
        segments:
          type: array
          items:
            $ref: '#/components/schemas/ContentSegment'
...
    DuaListResponse:
      type: object
      required: [data, pagination, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DuaSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    DuaResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/Dua'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    DuaEvidenceResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [recordId, canonicalUrl, revisionNumber, publishedAt, references, taxonomy, verificationHistory, corrections]
          properties:
            recordId:
              type: string
            canonicalUrl:
              type: string
              format: uri
            revisionNumber:
              type: integer
              minimum: 1
            publishedAt:
              type: string
              format: date-time
            collection:
              type: object
              nullable: true
            references:
              type: array
              items:
                type: object
            taxonomy:
              type: array
              items:
                type: object
            verificationHistory:
              type: array
              items:
                type: object
            corrections:
              type: array
              items:
                type: object
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    CollectionSummary:
      type: object
      required: [id, slug, contentType, title, recordCount, bookCount, chapterCount, verificationStatus]
      properties:
        id: { type: string }
        slug: { type: string }
        contentType: { type: string, enum: [dua, hadith] }
        title: { type: string }
        titleArabic: { type: string, nullable: true }
        recordCount: { type: integer, minimum: 0 }
        bookCount: { type: integer, minimum: 0 }
        chapterCount: { type: integer, minimum: 0 }
        verificationStatus: { type: string, enum: [verified] }
...
    HadithListResponse:
      type: object
      required: [data, pagination, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/HadithSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    AskRequest:
      type: object
      required: [question]
      properties:
        question:
          type: string
          minLength: 5
          maxLength: 500
...